Added navigation with each button leading to a page
I took out the splash screen; we could make our own later.
I also added a navigation bar with 4 buttons. each button leads to a different page.
The buttons are dynamic, so they are highlighted when you are on the page which a button is supposed to lead to.
So basicaly, if you are in the page for button 1, then button 1 will be highlighted.